requirements.txt
python-dotenv
langchain-core
langchain-google-genai
langchain-postgres
langgraph
google-generativeai
psycopg2-binary
pandas
pydantic
fastapi
uvicorn